Fix critical runtime issues: null logger deref, UB in empty query lists, dishonest INI parse errors
Problem:
Several error paths could crash the process or silently misbehave: a worker thread could dereference a null logger under a specific failure path, an empty query list could reach code that assumed a positive size, and a configuration parse error was silently discarded instead of being reported.

Cause:
These are longstanding gaps in error handling on paths that only trigger under failure conditions (exhausted retries, malformed or missing input), so they went unexercised under normal operation.

Solution:
Add the missing guards so these paths fail loudly and safely instead of crashing or silently continuing: fall back to a working logger, reject empty input explicitly, and propagate parse errors instead of discarding them. Also hardened related numeric parsing against overflow and removed an unintended side effect from a lookup helper.

+++ src/cIniReader.hpp
@@ -1,3 +1,4 @@
+#include <cstdint>
 #include <eTypes.hpp>
 #include <string>
 #include <unordered_map>
@@ -16,7 +17,8 @@ class INIReader {
   int ParseError() const { return _error; };
   std::vector<std::string> GetSections() const { return _sections; };
   std::string Get(std::string, std::string name, std::string default_value);
-  int GetInteger(std::string section, std::string name, int default_value);
+  int64_t GetInteger(std::string section, std::string name,
+                     int64_t default_value);
   bool GetBoolean(std::string section, std::string name, bool default_value);
   eDBTYPE getDbType(std::string section, std::string name,
                     eDBTYPE default_value);
